Starts building services for filters
