+++ ember-app/app/services/filters.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+var FiltersService = Ember.Service.extend({
+
+});
+
+export default FiltersService;
